+++ components/camel-redis/src/test/java/org/apache/camel/component/redis/processor/aggregate/integration/MyAggregationStrategy.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.redis.processor.aggregate;
+package org.apache.camel.component.redis.processor.aggregate.integration;
 
 import org.apache.camel.AggregationStrategy;
 import org.apache.camel.Exchange;
